Review comments implementation.
