+++ src/test/java/net/snowflake/client/jdbc/ConnectionWithDisableOCSPModeLatestIT.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024 Snowflake Computing Inc. All right reserved.
+ * Copyright (c) 2025 Snowflake Computing Inc. All right reserved.
  */
 package net.snowflake.client.jdbc;
 
@@ -7,7 +7,6 @@ import static org.hamcrest.CoreMatchers.anyOf;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.junit.jupiter.api.Assertions.fail;
 
 import java.sql.DriverManager;
 import java.sql.SQLException;
@@ -38,84 +37,73 @@ public class ConnectionWithDisableOCSPModeLatestIT extends BaseJDBCTest {
 
   /** Test connectivity with disableOCSPChecksMode and insecure mode enabled. */
   @Test
-  public void testDisableOCSPChecksModeAndInsecureMode() throws SQLException {
+  public void testDisableOCSPChecksModeAndInsecureModeSet() throws SQLException {
 
-    String deploymentUrl =
-        "jdbc:snowflake://sfcsupport.snowflakecomputing.com?disableOCSPChecks=true&insecureMode=true";
-    Properties properties = new Properties();
-
-    properties.put("user", "fakeuser");
-    properties.put("password", "fakepwd");
-    properties.put("account", "fakeaccount");
-    SQLException thrown =
-        assertThrows(
-            SQLException.class,
-            () -> {
-              DriverManager.getConnection(deploymentUrl, properties);
-            });
-
+    boolean disableOCSPChecks = true;
+    boolean insecureMode = true;
     assertThat(
-        thrown.getErrorCode(), anyOf(is(INVALID_CONNECTION_INFO_CODE), is(BAD_REQUEST_GS_CODE)));
-  }
-
-  /** Test connectivity with disableOCSPChecksMode enabled and insecure mode disabled. */
-  @Test
-  public void testDisableOCSPChecksModeAndInsecureModeMismatched() throws SQLException {
-
-    String deploymentUrl =
-        "jdbc:snowflake://sfcsupport.snowflakecomputing.com?disableOCSPChecks=true&insecureMode=false";
-    Properties properties = new Properties();
-
-    properties.put("user", "fakeuser");
-    properties.put("password", "fakepwd");
-    properties.put("account", "fakeaccount");
-    SQLException thrown =
-        assertThrows(
-            SQLException.class,
-            () -> {
-              DriverManager.getConnection(deploymentUrl, properties);
-            });
-
-    assertThat(thrown.getErrorCode(), anyOf(is(DISABLE_OCSP_INSECURE_MODE_MISMATCH)));
+        returnErrorCodeFromConnection(disableOCSPChecks, insecureMode),
+        anyOf(is(INVALID_CONNECTION_INFO_CODE), is(BAD_REQUEST_GS_CODE)));
   }
 
   /** Test production connectivity with only disableOCSPChecksMode enabled. */
   @Test
   public void testDisableOCSPChecksModeSet() throws SQLException {
-
-    String deploymentUrl =
-        "jdbc:snowflake://sfcsupport.snowflakecomputing.com?disableOCSPChecks=true";
-    Properties properties = new Properties();
-
-    properties.put("user", "fakeuser");
-    properties.put("password", "fakepwd");
-    properties.put("account", "fakeaccount");
-    SQLException thrown =
-        assertThrows(
-            SQLException.class,
-            () -> {
-              DriverManager.getConnection(deploymentUrl, properties);
-            });
-
+    boolean disableOCSPChecks = true;
     assertThat(
-        thrown.getErrorCode(), anyOf(is(INVALID_CONNECTION_INFO_CODE), is(BAD_REQUEST_GS_CODE)));
+        returnErrorCodeFromConnection(disableOCSPChecks, null),
+        anyOf(is(INVALID_CONNECTION_INFO_CODE), is(BAD_REQUEST_GS_CODE)));
   }
 
-  /** Test production connectivity with insecure mode enabled. */
+  /** Test production connectivity with only insecureMode enabled. */
   @Test
-  public void testEnableInsecureMode() throws SQLException {
+  public void testInsecureModeSet() throws SQLException {
+    boolean insecureMode = true;
+    assertThat(
+        returnErrorCodeFromConnection(null, insecureMode),
+        anyOf(is(INVALID_CONNECTION_INFO_CODE), is(BAD_REQUEST_GS_CODE)));
+  }
+
+  /** Test production connectivity with disableOCSPChecksMode enabled AND insecureMode disabled. */
+  @Test
+  public void testDisableOCSPChecksModeAndInsecureModeMismatched() throws SQLException {
+    boolean disableOCSPChecks = true;
+    boolean insecureMode = false;
+    assertThat(
+        returnErrorCodeFromConnection(disableOCSPChecks, insecureMode),
+        anyOf(is(DISABLE_OCSP_INSECURE_MODE_MISMATCH)));
+  }
+
+  /**
+   * Helper method to return error code from connection.
+   *
+   * @param disableOSCPChecks
+   * @param isInsecureMode
+   * @return SF Error code
+   * @throws SQLException
+   */
+  public int returnErrorCodeFromConnection(Boolean disableOSCPChecks, Boolean isInsecureMode)
+      throws SQLException {
     String deploymentUrl = "jdbc:snowflake://sfcsupport.snowflakecomputing.com?insecureMode=true";
     Properties properties = new Properties();
 
     properties.put("user", "fakeuser");
     properties.put("password", "fakepwd");
     properties.put("account", "fakeaccount");
-    try {
-      DriverManager.getConnection(deploymentUrl, properties);
-      fail();
-    } catch (SQLException e) {
-      assertThat(
-          e.getErrorCode(), anyOf(is(INVALID_CONNECTION_INFO_CODE), is(BAD_REQUEST_GS_CODE)));
+    if (disableOSCPChecks != null) {
+      properties.put("disableOCSPChecks", disableOSCPChecks);
     }
+    if (isInsecureMode != null) {
+      properties.put("insecureMode", isInsecureMode);
+    }
+
+    SQLException thrown =
+        assertThrows(
+            SQLException.class,
+            () -> {
+              DriverManager.getConnection(deploymentUrl, properties);
+            });
+
+    return (thrown.getErrorCode());
   }
 }
